Getting better all the time
* Better documentation for importing stackoverflow.com
* Adding Polly to have retries on SQL timeouts
* Improved progress bars for huge imports.

+++ src/Soddi/Services/IArchivedDataProcessor.cs
@@ -2,6 +2,6 @@
 
 public interface IArchivedDataProcessor
 {
-    IEnumerable<(string fileName, Stream stream, int size)> GetFiles();
+    IEnumerable<(string fileName, Stream stream, long size)> GetFiles();
     long GetTotalFileSize();
 }
